maps

+++ window.cpp
@@ -11,7 +11,7 @@ Window::Window() :
 
 }
 
-Window::Window(uint32_t w, uint32_t h, uint32_t x, uint32_t y) :
+Window::Window(int32_t w, int32_t h, int32_t x, int32_t y) :
 	win(newwin(h, w, x, y))
 {
 	if (this->win == 0) {
@@ -23,10 +23,10 @@ Window::Window(uint32_t w, uint32_t h, uint32_t x, uint32_t y) :
 Window::Window(Window &wwin) :
 	win(0)
 {
-	uint32_t x = wwin.PosX();
-	uint32_t y = wwin.PosY();
-	uint32_t w = wwin.Width();
-	uint32_t h = wwin.Height();
+	int32_t x = wwin.PosX();
+	int32_t y = wwin.PosY();
+	int32_t w = wwin.Width();
+	int32_t h = wwin.Height();
 	this->win = newwin(h, w, x, y);
 	if (this->win == 0) {
 		cerr << "Error while creating window instance: " << this 
@@ -50,9 +50,9 @@ void Window::KeyPad(bool bf)
 	}
 }
 
-uint32_t Window::GetChar()
+int32_t Window::GetChar()
 {
-	return uint32_t(wgetch(this->win));
+	return wgetch(this->win);
 }
 
 void Window::Refresh()
@@ -62,22 +62,22 @@ void Window::Refresh()
 	}
 }
 
-uint32_t Window::Width()
+int32_t Window::Width()
 {
 	return getmaxx(this->win);
 }
 
-uint32_t Window::Height()
+int32_t Window::Height()
 {
 	return getmaxy(this->win);
 }
 
-uint32_t Window::PosX()
+int32_t Window::PosX()
 {
 	return getbegx(this->win);
 }
 
-uint32_t Window::PosY()
+int32_t Window::PosY()
 {
 	return getbegy(this->win);
 }
